emosync-streamlit/requirements.txt
streamlit==1.28.0
ultralytics==8.0.202
python-dotenv
requests
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
Pillow==10.1.0
